Analysis of the conjugacy classes via invariants

AntiIdClosedGroup.GArf — Method

GArf(f::Function, V::Vector; readable=true)

Implements the GArf-invariant by applying the function f to all vectors in V (List of Vectors), see: $\text{GArf}(q) = phase\left\{ \sum_{\vec x\in V_g} i^{q(\vec x)} \right\}$

AntiIdClosedGroup.generalized_arf — Method

generalized_arf(g::SparseMatrixCSC{ℤ₂, Int64}, n::Int64; readable = true)

Combines the calculation of the kernel Vg = ker_g_plus_id(g,n) with the generalized Arf invariant GArf(quadratic_form, Vg). See ker_g_plus_id, quadratic_form and GArf.

AntiIdClosedGroup.ker_g_plus_id — Method

ker_g_plus_id(g::SparseMatrixCSC{ℤ₂, Int64}, n::Int64)::Vector{Vector{ℤ₂}}

We are interested in $Vg = \{\vec{x} \in \mathbb{Z}_2^n \, | \, g \cdot \vec{x} = \vec{x}\} = \{\vec{x} \in \mathbb{Z}_2^n \, | \, (g \pm \mathbf{1}) \cdot \vec{x} = \vec{0}\}$. This means, $Vg$ is equivalent to the kernel, $ker(g\pm I)$ ($\pm$ is equivalent in ℤ₂).

AntiIdClosedGroup.ord — Method

ord(g::SparseMatrixCSC{ℤ₂, Int64}, n::Int64)::Int64

Calculate the order of a given element g, with g^order = identity. Only works for elements with finite order ("order > 0").

AntiIdClosedGroup.quadratic_form — Method

quadratic_form(x::Vector{ℤ₂})::Int8

Quadratic form $q: \mathbb{Z}_2^n \rightarrow \mathbb{Z}_4,\, \vec{x} \mapsto \sum_i ( x_i \mod{4})$.

AntiIdClosedGroup.rank_of_power_plus_id — Method

rank_of_power_plus_id(g::Matrix{ℤ₂}, power::Int64, n)::Int64

Alias for rank(g^power + id(n)).

AntiIdClosedGroup.tr_of_power — Method

tr_of_power(g::SparseMatrixCSC{ℤ₂, Int64}, power::Int64)::Int64

Alias for the trace tr(g^power).

AntiIdClosedGroup.ℤ₂ⁿ — Method

ℤ₂ⁿ(n::Int64)::Vector{Vector{ℤ₂}}

Calculate all 2^n possible vectors in $\mathbb{Z}_{2}^n$.